mbedtls: Support the sanity test on all PSA targets
The Mbed TLS sanity test should support all PSA targets including Mbed
OS PSA and TF-M targets. This requires the test to be compatible with
the current version of the PSA API, and to achieve this we make the
following changes:
* Replace the deprecated PSA_KEY_USAGE_SIGN and PSA_KEY_USAGE_VERIFY
with the PSA_KEY_USAGE_SIGN_HASH and PSA_KEY_USAGE_VERIFY_HASH.
* Remove call to mbedtls_psa_crypto_free() which is not part of the
standard PSA API and unavailable on TF-M. Even though the freeing call
is now removed, it is still safe for the test to call psa_crypto_init()
multiple times which is allowed by the PSA Crypto specification.