Fixed Join Request retransmission timing (Interop test)
On interop test 1.2.2.4, Join Request retransmission is expected to be 6 s
+ worst case air transmission. This delay is to accommodate for
JoinAccept through RX2.

Increasing the downlink-preamble-length caused the rx_timeout of RX2
window to be increased by ~320ms (on RX2 window, symbol length is about 32 ms).

Adding this change, increase the change of getting test 1.2.2.4 to
succeed (test 1.2.2.4 is subset of test 1.2.2).