TARGET_STM: fix unterminated multi-packet USB transaction
Fix issue #15384 by sending a zero-length packet if the packet sent had
the same size as the endpoint buffer size. This is needed to comply with
the USB protocol which assumes an ongoing multi-packet USB transaction
otherwise.

STM32 HAL PCD handle parameters are used to avoid data duplication in RAM.

Signed-off-by: Daniel Starke <daniel-email@gmx.net>